us av library instead!
Use the "av" library instead, which "just makes sense" since ```faster-library``` already uses it anyways.  This removes the ```pyaudio``` dependency as well as that of ```scipy``` and ```numpy```.  It also obviates having to rely on users to know how to install ffmpeg since ```av``` is basically a wrapper for ffmpeg.